[FEAT] Refactor PlayScreen: Modularize components, implement GameCarousel and BeastPlayDisplay, and integrate mini-games data
